+++ sdk-extensions/declarative-config/src/main/java/io/opentelemetry/sdk/autoconfigure/declarativeconfig/model/internal/ExperimentalPrometheusMetricExporterModel.java
@@ -5,14 +5,10 @@
 
 package io.opentelemetry.sdk.autoconfigure.declarativeconfig.model.internal;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import com.fasterxml.jackson.annotation.JsonValue;
 import io.opentelemetry.sdk.autoconfigure.declarativeconfig.model.IncludeExcludeModel;
-import java.util.HashMap;
-import java.util.Map;
 import javax.annotation.Generated;
 import javax.annotation.Nullable;
 
@@ -25,7 +21,7 @@ import javax.annotation.Nullable;
   "resource_constant_labels",
   "translation_strategy"
 })
-@Generated("jsonschema2pojo")
+@Generated("io.opentelemetry.gradle.DeclarativeConfigPojoGenerator")
 public class ExperimentalPrometheusMetricExporterModel {
 
   @Nullable private String host;
@@ -33,10 +29,7 @@ public class ExperimentalPrometheusMetricExporterModel {
   @Nullable private Boolean scopeInfoEnabled;
   @Nullable private Boolean targetInfoEnabledDevelopment;
   @Nullable private IncludeExcludeModel resourceConstantLabels;
-
-  @Nullable
-  private ExperimentalPrometheusMetricExporterModel.ExperimentalPrometheusTranslationStrategy
-      translationStrategy;
+  @Nullable private ExperimentalPrometheusTranslationStrategyModel translationStrategy;
 
   /**
    * Configure host.
@@ -148,15 +141,13 @@ public class ExperimentalPrometheusMetricExporterModel {
    */
   @JsonProperty("translation_strategy")
   @Nullable
-  public ExperimentalPrometheusMetricExporterModel.ExperimentalPrometheusTranslationStrategy
-      getTranslationStrategy() {
+  public ExperimentalPrometheusTranslationStrategyModel getTranslationStrategy() {
     return translationStrategy;
   }
 
   @JsonProperty("translation_strategy")
   public ExperimentalPrometheusMetricExporterModel withTranslationStrategy(
-      ExperimentalPrometheusMetricExporterModel.ExperimentalPrometheusTranslationStrategy
-          translationStrategy) {
+      ExperimentalPrometheusTranslationStrategyModel translationStrategy) {
     this.translationStrategy = translationStrategy;
     return this;
   }
@@ -224,57 +215,5 @@ public class ExperimentalPrometheusMetricExporterModel {
               : this.translationStrategy.equals(that.translationStrategy));
     }
     return false;
-  }
-
-  @Generated("jsonschema2pojo")
-  public enum ExperimentalPrometheusTranslationStrategy {
-    UNDERSCORE_ESCAPING_WITH_SUFFIXES("underscore_escaping_with_suffixes"),
-    UNDERSCORE_ESCAPING_WITHOUT_SUFFIXES_DEVELOPMENT(
-        "underscore_escaping_without_suffixes/development"),
-    NO_UTF_8_ESCAPING_WITH_SUFFIXES_DEVELOPMENT("no_utf8_escaping_with_suffixes/development"),
-    NO_TRANSLATION_DEVELOPMENT("no_translation/development");
-    private final String value;
-    private static final Map<
-            String,
-            ExperimentalPrometheusMetricExporterModel.ExperimentalPrometheusTranslationStrategy>
-        CONSTANTS =
-            new HashMap<
-                String,
-                ExperimentalPrometheusMetricExporterModel
-                    .ExperimentalPrometheusTranslationStrategy>();
-
-    static {
-      for (ExperimentalPrometheusMetricExporterModel.ExperimentalPrometheusTranslationStrategy c :
-          values()) {
-        CONSTANTS.put(c.value, c);
-      }
-    }
-
-    ExperimentalPrometheusTranslationStrategy(String value) {
-      this.value = value;
-    }
-
-    @Override
-    public String toString() {
-      return this.value;
-    }
-
-    @JsonValue
-    public String value() {
-      return this.value;
-    }
-
-    @JsonCreator
-    public static ExperimentalPrometheusMetricExporterModel
-            .ExperimentalPrometheusTranslationStrategy
-        fromValue(String value) {
-      ExperimentalPrometheusMetricExporterModel.ExperimentalPrometheusTranslationStrategy constant =
-          CONSTANTS.get(value);
-      if (constant == null) {
-        throw new IllegalArgumentException(value);
-      } else {
-        return constant;
-      }
-    }
   }
 }